Restructured some header files Signed-off-by:doe300 <stadeldani@web.de>
