+++ include/sip/SDPMessageHandler.h
@@ -10,7 +10,7 @@
 
 #include "configuration.h"
 #include "KeyValuePairs.h"
-#include "rtp/RTPPackageHandler.h"
+#include "rtp/RTPHeader.h"
 #include "sip/SIPUserAgent.h"
 #include "sip/SupportedFormats.h"
