+++ src/sip/SupportedFormats.cpp
@@ -1,7 +1,6 @@
 
 #include "sip/SupportedFormats.h"
 #include "AudioProcessorFactory.h"
-#include "rtp/RTPPackageHandler.h"
 
 #include <iostream>
